requirements.txt
transformers
torch
faiss-cpu
PyPDF2
sentence-transformers
streamlit
markdown